Kernel Sources from QNAP
Downloaded from https://sourceforge.net/projects/qosgpl

$ sha256sum GPL_QTS-4.3.x-20161219_Kernel.tar.gz
cc72e6debc6741c41d84e7021952ee60262a74179f9d89e60c8deb06f4565b79
GPL_QTS-4.3.x-20161219_Kernel.tar.gz

GPL_TS/src/linux-3.10.20

+++ drivers/block/loop.c
@@ -344,8 +344,8 @@ lo_splice_actor(struct pipe_inode_info *pipe, struct pipe_buffer *buf,
 		size = p->bsize;
 
 	if (lo_do_transfer(lo, READ, page, buf->offset, p->page, p->offset, size, IV)) {
-		printk(KERN_ERR "loop: transfer error block %ld\n",
-		       page->index);
+		printk(KERN_ERR "loop: transfer error block %lld\n",
+		       (unsigned long long)page->index);
 		size = -EINVAL;
 	}
